fix(all/lunaranime): parse the flight payload with the core Next.js helper
The hand-rolled `self.__next_f.push` regex matched the JS string literal with
an alternation inside a quantifier, which java.util.regex expands recursively:
one frame per character. The reader page ships the whole payload in a single
44KB push, so extractSeeds blew the stack before it read a byte.

extractNextJs already parses the flight rows and walks them, so the regex and
the flat-object scan both go away. The seed props move to a LunarSeedPropsDto
with a map-delegating serializer, since the site randomises the prop names per
response and only the decoded payload tells the real set from the decoys.
